+++ MyFirstJavaProject/src/com/walgreens/examples/SecondMathOperation.java
@@ -1,0 +1,21 @@
+package com.walgreens.examples;
+
+public class SecondMathOperation {
+
+	public static void main(String[] args){
+		
+			addition();
+	}
+	
+	  public static void addition(){
+		 //  syntex; datatype variable = value
+	     // example of datatype; int double;
+	        int number1 =10;
+	        int number2 = 20;
+	        int number3 = 10;
+	        
+	        int result = number1 + number2 + number3;
+	       System.out.println("Addition is : "+result);
+	       
+	   }
+}
