+++ MyFirstJavaProject/src/com/walgreens/examples/ThirdMathOperation.java
@@ -1,0 +1,25 @@
+package com.walgreens.examples;
+
+public class ThirdMathOperation {
+
+	public static  void main (String[]args){
+		
+		addSub();
+		
+	}
+	public static  void addSub(){
+	
+	//Syntex :  Datatype variable = value
+	
+	//Example of datatype:  int, double
+	
+		int number1 = 20;
+		int number2 = 23;
+		int number3 = 32;
+		
+		int result = number1 + number2 - number3;
+		System.out.println("Answer is : "+ result);
+		
+	}
+	
+	}
