--- MyFirstJavaProject/src/com/walgreens/examples/mysecondjavmathoperation.java
@@ -1,33 +1,0 @@
-package com.walgreens.examples;
-
-public class mysecondjavmathoperation {
-
-	public static void main(String[] args){
-		
-	
-			
-			//Addition();
-	
-
-	}
-	{
-         //addition logic
-	
-	// this method perform addition and subtraction logic
-	   
-	  // public static void addition and subtraction(){
-		 //  syntex; datatype variable = value
-	     // example of datatype; int double;
-	        int number1 =10;
-	        int number2 = 20;
-	        int number3 = 10;
-	        
-	        int result = number1 + number2+number3;
-	       System.out.println("result!!!!+result");
-	       
-	        
-	        
-	        
-	
-	   }
-}
