--- MyFirstJavaProject/src/com/walgreens/examples/mythirdjavamathoperation.java
@@ -1,33 +1,0 @@
-package com.walgreens.examples;
-
-public class mythirdjavamathoperation {
-
-	public static  void main (String[]args){
-		
-	
-		additionsubtraction();
-		
-	}
-	
-	// additionsubtraction logic
-		
-	
-	
-	// this method performs additionsubtraction logic
-	
-	public static  void additionsubtraction(){
-	
-	//Syntex; Datatype variable = value
-	
-	//Example of datatype; int, double
-	
-		int number1 = 10;
-		int number2 = 20;
-		int number3 = 30;
-		
-		int result = number1 + number2 -number3;
-		System.out.println("result####"+ result);
-		
-	}
-	
-	}
